thamer/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
alembic
python-dotenv
asyncpg
psycopg2-binary
pydantic_settings
aioredis
redis
passlib
python-jose
pydantic[email]
greenlet
boto3
jinja2
requests
python-multipart